+++ Source/Frontend/UI/Forms/BlastGeneratorForm.cs
@@ -75,13 +75,13 @@ namespace RTCV.UI
             //For some godforsaken reason, xmlSerializer deserialization wont fill this in as a bool so just use a string god help us all
             (dgvBlastGenerator.Columns["dgvEnabled"]).ValueType = typeof(string);
 
-            dgvInterval.Maximum = int.MaxValue;
-            dgvStartAddress.Maximum = int.MaxValue;
-            dgvEndAddress.Maximum = int.MaxValue;
-            dgvParam1.Maximum = int.MaxValue;
-            dgvParam2.Maximum = int.MaxValue;
-            dgvLifetime.Maximum = int.MaxValue;
-            dgvExecuteFrame.Maximum = int.MaxValue;
+            dgvInterval.Maximum = decimal.MaxValue;
+            dgvStartAddress.Maximum = decimal.MaxValue;
+            dgvEndAddress.Maximum = decimal.MaxValue;
+            dgvParam1.Maximum = decimal.MaxValue;
+            dgvParam2.Maximum = decimal.MaxValue;
+            dgvLifetime.Maximum = decimal.MaxValue;
+            dgvExecuteFrame.Maximum = decimal.MaxValue;
 
             Colors.SetRTCColor(Colors.GeneralColor, this);
         }
